Book and category CRUD

+++ database/migrations/2024_05_16_105801_create_books_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('category_id')->constrained('categories')->onDelete('cascade');
             $table->string('name');
             $table->text('description');
-            $table->decimal('price', 8, 2);
+            $table->unsignedInteger('price');
             $table->string('author');
             $table->timestamps();
         });
